Added exchange service with test implementation

+++ src/main/java/com/revolut/money/delivery/controller/AccountManagementController.java
@@ -11,6 +11,8 @@ import io.vertx.core.json.Json;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.RoutingContext;
 
+import java.math.BigDecimal;
+
 
 public class AccountManagementController {
 
@@ -41,7 +43,7 @@ public class AccountManagementController {
     public void createAccount(RoutingContext context) {
         JsonObject bodyAsJson = context.getBodyAsJson();
         String actHolder = bodyAsJson.getString("holder");
-        double amount = bodyAsJson.getDouble("amount");
+        BigDecimal amount = BigDecimal.valueOf(bodyAsJson.getDouble("amount"));
         String currencyCode = bodyAsJson.getString("currencyCode");
 
         Money money = new Money(amount, currencyCode);
@@ -88,7 +90,7 @@ public class AccountManagementController {
         String accountHolder = bodyAsJson.getString("accountHolder");
         int accountNum = bodyAsJson.getInteger("accountNum");
         AccountId accountId = new AccountId(accountHolder, accountNum);
-        double amount = bodyAsJson.getDouble("amount");
+        BigDecimal amount = BigDecimal.valueOf(bodyAsJson.getDouble("amount"));
         String currency = bodyAsJson.getString("currency");
         Money moneyToPut = new Money(amount, currency);
         transactionService.deposit(accountId, moneyToPut);
@@ -100,7 +102,7 @@ public class AccountManagementController {
         String accountHolder = bodyAsJson.getString("accountHolder");
         int accountNum = bodyAsJson.getInteger("accountNum");
         AccountId accountId = new AccountId(accountHolder, accountNum);
-        double amount = bodyAsJson.getDouble("amount");
+        BigDecimal amount = BigDecimal.valueOf(bodyAsJson.getDouble("amount"));
         String currency = bodyAsJson.getString("currency");
         Money moneyToFetch = new Money(amount, currency);
         transactionService.withdraw(accountId, moneyToFetch);
@@ -118,7 +120,7 @@ public class AccountManagementController {
         int toAccountNum = bodyAsJson.getInteger("toAccountNum");
         AccountId toAccountId = new AccountId(toAccountHolder, toAccountNum);
 
-        double amount = bodyAsJson.getDouble("amount");
+        BigDecimal amount = BigDecimal.valueOf(bodyAsJson.getDouble("amount"));
         String currency = bodyAsJson.getString("currency");
         Money moneyToTransfer = new Money(amount, currency);
